Add DiscoveryUtils
Handle discovery API listings query parameter formatting in a utility
module.
